Put FakeRummageableIndex in own file
Changes to Ruby or Rails, means that classes must be in a file with a
name that matches the class name, otherwise a TypeError is thrown:

```
TypeError (superclass mismatch for class FakeRummageableIndex)
```
